www: render blog mdx on server
