+++ apps/www/app/blog/[slug]/page.tsx
@@ -89,14 +89,17 @@ export default async function BlogPostPage({ params }: { params: Promise<Params>
   const { isEnabled: isDraft } = await draftMode()
 
   const matter = (await import('gray-matter')).default
-  const { mdxSerialize } = await import('lib/mdx/mdxSerialize')
 
   try {
     const postContent = await getPostdata(slug, '_blog')
     const parsedContent = matter(postContent) as unknown as MatterReturn
     const content = parsedContent.content
     const tocDepth = (parsedContent.data as any)?.toc_depth ?? 3
-    const mdxSource = await mdxSerialize(content, { tocDepth })
+    const { preprocessMdxWithCodeTabs } = await import('~/components/CodeTabs')
+    const { addSelfClosingTags } = await import('lib/mdx/addSelfClosingTags')
+    const { extractToc } = await import('lib/mdx/extractToc')
+    const preprocessed = await preprocessMdxWithCodeTabs(addSelfClosingTags(content))
+    const tocResult = await extractToc(preprocessed, tocDepth)
     const { generateReadingTime } = await import('lib/helpers')
     const blogPost = {
       ...parsedContent.data,
@@ -113,12 +116,15 @@ export default async function BlogPostPage({ params }: { params: Promise<Params>
     const currentIndex = allPosts.findIndex((post) => post.slug === slug)
     const nextPost = currentIndex === allPosts.length - 1 ? null : allPosts[currentIndex + 1]
     const prevPost = currentIndex === 0 ? null : allPosts[currentIndex - 1]
-    const tocResult = (mdxSource as any).scope?.toc || { content: '' }
-    const processedContent = tocResult.content
+    const frontmatterTags = (parsedContent.data as { tags?: Array<string | { name: string }> })
+      ?.tags
+    const tagNames = Array.isArray(frontmatterTags)
+      ? frontmatterTags.map((t) => (typeof t === 'string' ? t : t?.name)).filter(Boolean)
+      : undefined
     const relatedPosts = getSortedPosts({
       directory: '_blog',
       limit: 3,
-      tags: (mdxSource as { scope: { tags?: string[] } }).scope.tags,
+      tags: tagNames,
       currentPostSlug: slug,
     }) as unknown as (BlogData & PostReturnType)[]
 
@@ -129,15 +135,14 @@ export default async function BlogPostPage({ params }: { params: Promise<Params>
       blog: {
         ...(blogPost as any),
         slug,
-        content: mdxSource,
-        toc: {
-          ...tocResult,
-          content: processedContent,
-        },
+        content: preprocessed,
+        toc: tocResult,
       } as any,
       isDraftMode: isDraft,
     }
 
     return <BlogPostClient {...props} />
-  } catch {}
+  } catch (err) {
+    throw err
+  }
 }
