+++ apps/www/types/post.ts
@@ -1,5 +1,3 @@
-import type { MDXRemoteSerializeResult } from 'next-mdx-remote'
-
 import type { getSortedPosts } from '@/lib/posts'
 
 export interface PostTypes {
@@ -78,7 +76,7 @@ export type Blog = {
   slug: string
   title: string
   description?: string
-  content: MDXRemoteSerializeResult
+  content: string
   toc: string | { content: string }
   author?: string
   authors?: StaticAuthor[]
@@ -113,7 +111,7 @@ export type BlogData = {
   slug: string
   title: string
   description?: string
-  content: MDXRemoteSerializeResult
+  content: string
   toc: string | { content: string }
   author?: string
   authors?: StaticAuthor[]
